packages/bot/requirements.txt
web3==7.4.0
python-dotenv
